refactor(resolution): single skillset-resolution edge; qualify public everywhere
Consolidate scattered skillset name-resolution logic into one edge and make
naming consistent across the CLI.

- Add src/cli/skillsetResolution.ts as the single resolution edge
  (resolveUserSkillsetRef, namespaceCreateSkillsetName, canonicalSkillsetName),
  moved out of norijson/skillset.ts which now holds only path/discovery
  primitives (+ new skillsetPath helper). defaultOrg is applied once, at the edge.
- Retire the 'public = bare' convention: namespacedName always returns
  <org>/<name> (public -> public/<name>); collapse namespacedOnDiskName into it;
  remove dead localSkillsetName. Skills/subagents/search display qualify too.
- Remove resolveSwitchSkillsetName; the switchSkillset op no longer re-applies
  defaultOrg and never re-resolves an already-qualified name.
- Fix: 'install public/foo' with a configured defaultOrg previously threw
  'Profile not found' (op re-namespaced the collapsed bare name); the qualified
  name now flows through unchanged.
- Replace inline path.relative identity re-derivations with skillsetIdentity.
