+++ src/cli/commands/registry-download/docs.md
@@ -12,7 +12,7 @@ Registered via `@/src/cli/commands/noriSkillsetsCommands.ts` as the `download` c
 
 ### Core Implementation
 
-`registryDownloadMain` loads config **before** parsing the spec so a bare (non-namespaced) name can resolve against the configured `defaultOrg` via `parseNamespacedPackage` (from @/src/utils/url.ts); the `defaultOrg` is passed only when no explicit `--registry` was given, and when a bare name is thereby routed to a non-public org the command logs `formatDefaultOrgNotice` so the redirect is visible. It then follows a callback-driven flow pattern using `registryDownloadFlow`. The `onSearch` callback resolves which registry to use (explicit `--registry`, namespaced `org/package`, or public fallback), fetches the packument, compares versions against any locally installed copy, and returns a status (`ready`, `already-current`, `list-versions`, or `error`). The `onDownload` callback fetches the tarball, installs it via the atomic-replacement primitives from @/src/packaging/atomicReplace.ts, writes provenance via `writeVersionInfo` from @/src/packaging/provenance.ts, and resolves both skill and subagent dependencies from the skillset's `nori.json`. The install target is chosen via `namespacedOnDiskName` from @/src/utils/url.ts, so public packages are written to the `public/` storage bucket (`profiles/public/<name>`) and org packages to `profiles/<orgId>/<name>` (see the storage-bucket model in @/src/norijson/docs.md).
+`registryDownloadMain` loads config **before** parsing the spec so a bare (non-namespaced) name can resolve against the configured `defaultOrg` via `parseNamespacedPackage` (from @/src/utils/url.ts); the `defaultOrg` is passed only when no explicit `--registry` was given, and when a bare name is thereby routed to a non-public org the command logs `formatDefaultOrgNotice` so the redirect is visible. It then follows a callback-driven flow pattern using `registryDownloadFlow`. The `onSearch` callback resolves which registry to use (explicit `--registry`, namespaced `org/package`, or public fallback), fetches the packument, compares versions against any locally installed copy, and returns a status (`ready`, `already-current`, `list-versions`, or `error`). The `onDownload` callback fetches the tarball, installs it via the atomic-replacement primitives from @/src/packaging/atomicReplace.ts, writes provenance via `writeVersionInfo` from @/src/packaging/provenance.ts, and resolves both skill and subagent dependencies from the skillset's `nori.json`. The install target is chosen by splitting `namespacedName` from @/src/utils/url.ts on `/`, so public packages are written to the `public/` storage bucket (`profiles/public/<name>`) and org packages to `profiles/<orgId>/<name>` (see the storage-bucket model in @/src/norijson/docs.md). `namespacedName` is now the single qualified name — the former separate `namespacedOnDiskName` was collapsed into it once the retired "public = bare" convention made display, disk, and identity all agree.
 
 Registry search supports three strategies: explicit registry URL, namespace-derived registry, and public registry. For namespaced packages under unified auth, the org registry URL, membership check, and token acquisition come from `resolveOrgRegistryAuth` in @/src/core/registryAuthResolution.ts; the all-org parallel search and anonymous public paths still build URLs directly via `buildOrganizationRegistryUrl`. The explicit-`registryUrl` branch (used by switch's provenance-pinned re-download) resolves its auth token the same way: it **prefers unified org auth via `resolveOrgRegistryAuth`** when the pinned URL is an org registry the user belongs to (org id extracted via `extractOrgId` from @/src/utils/url.ts, membership taken from `config.auth.organizations`), and only falls back to `getRegistryAuth` for the home org and local dev. This makes the pinned-registry path authenticate identically to the derived-registry path — both the packument search and the tarball download — so a multi-org member whose home org differs from the pinned org still authenticates. When no namespace or registry is specified and the user is authenticated, it searches all configured organization registries in parallel before falling back to public. Per-registry lookup and the version-list / multiple-matches messages use `searchSpecificRegistry`, `formatVersionList`, and `formatMultipleMatchesError` from @/src/packaging/registryLookup.ts, with `registrarApi` fetches and auth-token resolution injected as callbacks. A bare name routed through `defaultOrg` becomes org-namespaced, so it resolves to that single org registry with no public fallback on a miss (dependency-confusion safety). Namespace packages that require auth but the user lacks access produce specific error messages.
