+++ src/cli/commands/registry-upload/docs.md
@@ -14,7 +14,7 @@ Path: @/src/cli/commands/registry-upload
 - Uses `@/api/registrar.js` (`registrarApi.uploadSkillset`, `registrarApi.getPackument`) for registry communication
 - Auth tokens are obtained via `@/api/registryAuth.js` (`getRegistryAuthToken`) using refresh tokens, broker-managed direct ID tokens, saved API tokens, or a matching `NORI_API_TOKEN`; credential predicates (`hasRegistryAuthCredentials`, `toRegistryAuth`) come from @/src/api/authCredentials.ts, and the org-scoped membership/URL/auth resolution is shared with the download commands via `resolveOrgRegistryAuth` from @/src/core/registryAuthResolution.ts
 - Skillset metadata is read/written through `@/norijson/nori.js` (`readSkillsetMetadata`, `writeSkillsetMetadata`)
-- The local skillset directory is located via `resolveSkillsetDir` from `@/norijson/skillset.js`, so a bare name resolves across the `personal/`/`public/` storage buckets and the legacy flat location (see the storage-bucket model in @/src/norijson/docs.md); a name that resolves nowhere fails with "Skillset not found"
+- The local skillset *source* is located via `resolveSkillsetDir` from `@/norijson/skillset.js`, using **bucket precedence for a public package**: a public target is looked up by its bare `packageName` (so `resolveSkillsetDir` walks `personal/` -> `public/` -> legacy flat), while an org target is looked up at its `<org>/<name>` namespace. This lets a locally-created `personal/` or legacy-flat skillset still be published to public. The source lookup is deliberately separate from the registry *target*, which is derived from the parsed namespace and displayed fully qualified via `namespacedName` (public = `public/<name>`); a name that resolves nowhere fails with "Skillset not found" (see the storage-bucket model in @/src/norijson/docs.md)
 - The interactive upload flow is driven by `uploadFlow` and `listVersionsFlow` from `@/cli/prompts/flows/`
 - Namespace parsing uses `parseNamespacedPackage` from `@/utils/url.js`, shared with `registryDownload` and `skillDownload`
