+++ src/cli/commands/skill-download/docs.md
@@ -22,6 +22,6 @@ The `onDownload` callback installs the tarball via the atomic-replacement primit
 
 ### Things to Know
 
-The `--skillset` flag lets the user target a specific skillset for manifest updates; otherwise it defaults to the active skillset from config. A user-supplied `--skillset` is resolved via `resolveUserSkillsetRef` from @/src/norijson/skillset.ts (which resolves a bare name across the `personal/`/`public/` storage buckets and emits a one-time bare-name deprecation warning, suppressed under `--non-interactive`); the active-skillset fallback uses `resolveSkillsetDir` (no warning). The command tracks the resolved directory (`targetSkillsetDir`) separately from the display name (`targetSkillset`) — persistence writes go to the resolved directory while messages show the name. Like registry-download, updates use the atomic swap with backup/restore on failure from @/src/packaging/atomicReplace.ts, and failed fresh installs clean up their partial directory. The `--registry` flag and namespace prefix (`org/`) are mutually exclusive since the namespace implicitly determines the registry URL. The `nonInteractive` and `silent` params are threaded from the CLI registration layer through `skillDownloadMain` to `skillDownloadFlow`, where they control whether the "Re-download from registry?" confirm prompt is skipped when the skill is already at the current version. The coercion `nonInteractive ?? silent ?? false` is applied so `--silent` implies non-interactive behavior.
+The `--skillset` flag lets the user target a specific skillset for manifest updates; otherwise it defaults to the active skillset from config. A user-supplied `--skillset` is resolved via `resolveUserSkillsetRef` from @/src/cli/skillsetResolution.ts (which resolves a bare name across the `personal/`/`public/` storage buckets and emits a one-time bare-name deprecation warning, suppressed under `--non-interactive`); the active-skillset fallback uses `resolveSkillsetDir` from @/src/norijson/skillset.ts (no warning). The command tracks the resolved directory (`targetSkillsetDir`) separately from the display name (`targetSkillset`) — persistence writes go to the resolved directory while messages show the name. Like registry-download, updates use the atomic swap with backup/restore on failure from @/src/packaging/atomicReplace.ts, and failed fresh installs clean up their partial directory. The `--registry` flag and namespace prefix (`org/`) are mutually exclusive since the namespace implicitly determines the registry URL. The `nonInteractive` and `silent` params are threaded from the CLI registration layer through `skillDownloadMain` to `skillDownloadFlow`, where they control whether the "Re-download from registry?" confirm prompt is skipped when the skill is already at the current version. The coercion `nonInteractive ?? silent ?? false` is applied so `--silent` implies non-interactive behavior.
 
 Created and maintained by Nori.
