+++ src/cli/features/docs.md
@@ -86,7 +86,7 @@ The `--agent` global CLI option determines which agent implementation is used; t
 - `isInstalledAtDir`: True when `getInstalledSkillsetName` finds a marker, with a fallback to checking the agent's instructions file for `NORI-AI MANAGED BLOCK`.
 - `markInstall`: Writes `.nori-managed` marker containing the skillset name.
 - `installSkillset`: Parses the active skillset via `parseSkillset()`, backs up any external settings files declared by `agent.getExternalSettingsFiles()` via `backupSettingsFile()` from `@/src/cli/features/settingsBackup.ts` (idempotent; skips if backup already exists), runs all loaders from `agent.getLoaders()`, collects settings labels for a consolidated output note, writes the (agent, install dir) manifest (non-fatal on failure), emits a Skills note, and surfaces missing required environment variables via `checkRequiredEnv` from @/src/cli/features/envCheck.ts. It does **not** write `.nori-managed` markers; that responsibility belongs solely to `initMain` in @/src/cli/commands/init/init.ts, which calls `markInstall` for all default agents.
-- `switchSkillset`: Validates the target skillset exists (has `nori.json`) and logs success. Does not persist config. Resolves the target directory for a bare name across storage buckets via `resolveSkillsetDir` from @/src/norijson/skillset.ts (falling back to the legacy flat path).
+- `switchSkillset`: Validates the target skillset exists (has `nori.json`) and logs success under its namespaced identity. Does not persist config, and — by the single-resolution-edge invariant — **never applies `defaultOrg` or re-resolves the reference**: the name arrives already resolved at the command edge (see @/src/cli/skillsetResolution.ts), so the op only locates it on disk via `resolveSkillsetDir` (bucket precedence for a bare legacy name, direct for a namespaced one, falling back to `skillsetPath` when nothing exists yet) and displays `skillsetIdentity` (both from @/src/norijson/skillset.ts). The former `resolveSwitchSkillsetName` helper that re-loaded config and re-applied `defaultOrg` here has been removed.
 - `removeSkillset`: Runs `removeManagedFiles()` against the keyed (agent, install dir) manifest and both legacy manifest locations -- the old per-agent path from `getLegacyAgentManifestPath` and the agent-declared pre-manifest path when `getLegacyManifestPath` is present (declared as data on the claude-code table row -- shared code never branches on agent name). It then iterates over all loaders calling each loader's optional `uninstall()` method for non-settings cleanup, and restores external settings files from their pre-install backups via `restoreSettingsFile()` from `@/src/cli/features/settingsBackup.ts`. This full-file restore replaces the old approach of per-loader surgical key removal from settings files.
 - `detectLocalChanges`: Returns null immediately when `isInstalledAtDir` is false -- a directory Nori never installed to has no local changes by definition, and this guard keeps legacy pre-keying manifests (which describe whichever directory was installed last) from producing phantom changes elsewhere. Otherwise reads the keyed manifest (falling back through the legacy paths above) and compares file hashes.
 - `detectExistingConfig`: Scans the agent directory for instructions file, skills, subagents, and slashcommands using the agent's path getters.
@@ -116,7 +116,7 @@ The `--agent` global CLI option determines which agent implementation is used; t
 
 **Managed vs unmanaged boundary**: Top-level entries whose names start with `.` inside an agent's managed dirs (e.g., `<agent>/skills/.system/`) are treated as external, agent-owned content. They are invisible to `compareManifest` (so they never appear in the `switch` flow's "Local Changes Detected" prompt) and are skipped by `resetManagedDir` (so they survive reinstalls). This keeps Nori's reach scoped to non-dotfile children of declared `managedDirs`. See `manifest.ts` (`collectFiles` skips dotfile entries at every recursion level) and `shared/managedDirOps.ts`.
 
-Skillset path utilities (`getNoriDir`, `getNoriSkillsetsDir`), the `Skillset` type, `parseSkillset()`, and `listSkillsets()` now live in @/src/norijson/skillset.ts. Metadata CRUD functions (`readSkillsetMetadata`, `writeSkillsetMetadata`, `addSkillToNoriJson`, `ensureNoriJson`) now live in @/src/norijson/nori.ts.
+Skillset path utilities (`getNoriDir`, `getNoriSkillsetsDir`, `skillsetPath`), the `Skillset` type, `parseSkillset()`, and `listSkillsets()` now live in @/src/norijson/skillset.ts. Metadata CRUD functions (`readSkillsetMetadata`, `writeSkillsetMetadata`, `addSkillToNoriJson`, `ensureNoriJson`) now live in @/src/norijson/nori.ts. User-facing name resolution (`resolveUserSkillsetRef`, `namespaceCreateSkillsetName`, `canonicalSkillsetName`) lives above the commands in @/src/cli/skillsetResolution.ts; the operations here (e.g. `switchSkillset`) receive already-resolved identities and never re-apply `defaultOrg`.
 
 ### Things to Know
