+++ src/norijson/skillset.test.ts
@@ -22,9 +22,7 @@ import {
   getNoriSkillsetsDir,
   parseSkillset,
   listSkillsets,
-  namespaceCreateSkillsetName,
   resolveSkillsetDir,
-  resolveUserSkillsetRef,
 } from "@/norijson/skillset.js";
 
 describe("Shared Nori paths", () => {
@@ -218,20 +216,6 @@ describe("parseSkillset", () => {
       const resolved = await resolveSkillsetDir({ name: "acme/sessions" });
 
       expect(resolved).toBe(orgDir);
-    });
-
-    it("returns the namespaced identity for a bare name that lives in a bucket", async () => {
-      const publicDir = path.join(profilesDir, "public", "senior-swe");
-      await fs.mkdir(publicDir, { recursive: true });
-      await fs.writeFile(
-        path.join(publicDir, "nori.json"),
-        JSON.stringify({ name: "senior-swe", version: "1.0.0" }),
-      );
-
-      const resolved = await resolveUserSkillsetRef({ name: "senior-swe" });
-
-      expect(resolved?.dir).toBe(publicDir);
-      expect(resolved?.identity).toBe("public/senior-swe");
     });
 
     it("does not resolve a bare bucket name to the bucket directory", async () => {
@@ -664,153 +648,5 @@ describe("listSkillsets", () => {
 
     expect(profiles).toContain("personal/shared");
     expect(profiles).toContain("shared");
-  });
-});
-
-describe("namespaceCreateSkillsetName", () => {
-  it("prefixes a bare name with the default org", () => {
-    expect(
-      namespaceCreateSkillsetName({ name: "amol", defaultOrg: "myorg" }),
-    ).toBe("myorg/amol");
-  });
-
-  it("returns a bare name unchanged when there is no default org", () => {
-    expect(namespaceCreateSkillsetName({ name: "amol" })).toBe("amol");
-    expect(namespaceCreateSkillsetName({ name: "amol", defaultOrg: "" })).toBe(
-      "amol",
-    );
-  });
-
-  it("leaves an already-namespaced name unchanged", () => {
-    expect(
-      namespaceCreateSkillsetName({
-        name: "otherorg/amol",
-        defaultOrg: "myorg",
-      }),
-    ).toBe("otherorg/amol");
-  });
-
-  it("leaves an explicit bucket name unchanged", () => {
-    expect(
-      namespaceCreateSkillsetName({ name: "public/amol", defaultOrg: "myorg" }),
-    ).toBe("public/amol");
-    expect(
-      namespaceCreateSkillsetName({
-        name: "personal/amol",
-        defaultOrg: "myorg",
-      }),
-    ).toBe("personal/amol");
-  });
-});
-
-describe("resolveUserSkillsetRef deprecation warning", () => {
-  let testHomeDir: string;
-  let profilesDir: string;
-  let stderrOutput: Array<string>;
-
-  const seedBucket = async (bucket: string, name: string): Promise<void> => {
-    const dir = path.join(profilesDir, bucket, name);
-    await fs.mkdir(dir, { recursive: true });
-    await fs.writeFile(
-      path.join(dir, "nori.json"),
-      JSON.stringify({ name, version: "1.0.0" }),
-    );
-  };
-
-  beforeEach(async () => {
-    testHomeDir = await fs.mkdtemp(path.join(os.tmpdir(), "resolve-warn-"));
-    vi.mocked(os.homedir).mockReturnValue(testHomeDir);
-    profilesDir = path.join(testHomeDir, ".nori", "profiles");
-    await fs.mkdir(profilesDir, { recursive: true });
-    stderrOutput = [];
-    vi.spyOn(process.stderr, "write").mockImplementation(((
-      chunk: string | Uint8Array,
-    ): boolean => {
-      stderrOutput.push(String(chunk));
-      return true;
-    }) as typeof process.stderr.write);
-  });
-
-  afterEach(async () => {
-    vi.restoreAllMocks();
-    await fs.rm(testHomeDir, { recursive: true, force: true });
-    vi.clearAllMocks();
-  });
-
-  it("warns that a bare name is deprecated when it resolves to a bucket", async () => {
-    await seedBucket("public", "warn-alpha");
-
-    await resolveUserSkillsetRef({ name: "warn-alpha" });
-
-    const output = stderrOutput.join("");
-    expect(output).toContain("public/warn-alpha");
-    expect(output.toLowerCase()).toContain("deprecated");
-  });
-
-  it("prefers defaultOrg over bucketed local skillsets for bare names", async () => {
-    await seedBucket("dev", "high-autonomy");
-    await seedBucket("personal", "high-autonomy");
-    await seedBucket("public", "high-autonomy");
-
-    const resolved = await resolveUserSkillsetRef({
-      name: "high-autonomy",
-      defaultOrg: "dev",
-    });
-
-    expect(resolved?.identity).toBe("dev/high-autonomy");
-  });
-
-  it("does not apply defaultOrg to active-skillset fallbacks", async () => {
-    await seedBucket("dev", "high-autonomy");
-    await seedBucket("personal", "high-autonomy");
-
-    const resolved = await resolveUserSkillsetRef({
-      name: null,
-      activeSkillset: "high-autonomy",
-      defaultOrg: "dev",
-    });
-
-    expect(resolved?.identity).toBe("personal/high-autonomy");
-  });
-
-  it("keeps explicit public names scoped to public when defaultOrg is configured", async () => {
-    await seedBucket("dev", "high-autonomy");
-    await seedBucket("public", "high-autonomy");
-
-    const resolved = await resolveUserSkillsetRef({
-      name: "public/high-autonomy",
-      defaultOrg: "dev",
-    });
-
-    expect(resolved?.identity).toBe("public/high-autonomy");
-  });
-
-  it("does not fall back to a bucketed local skillset when the defaultOrg target is absent", async () => {
-    await seedBucket("personal", "high-autonomy");
-
-    const resolved = await resolveUserSkillsetRef({
-      name: "high-autonomy",
-      defaultOrg: "dev",
-    });
-
-    // Strict: a bare name with a default org resolves to <defaultOrg>/name only,
-    // never a same-named public/personal skillset.
-    expect(resolved).toBeNull();
-  });
-
-  it("does not warn when the namespaced identity is used", async () => {
-    await seedBucket("public", "warn-beta");
-
-    await resolveUserSkillsetRef({ name: "public/warn-beta" });
-
-    expect(stderrOutput).toHaveLength(0);
-  });
-
-  it("does not warn when the warning is suppressed", async () => {
-    await seedBucket("public", "warn-gamma");
-
-    await resolveUserSkillsetRef({ name: "warn-gamma", warn: false });
-
-    expect(stderrOutput).toHaveLength(0);
   });
 });
