+++ src/utils/docs.md
@@ -14,14 +14,11 @@ Every layer of the application imports from this module. The API layer uses `url
 
 **`url.ts`** handles URL normalization (preventing double slashes), org ID validation/extraction, and URL construction for two Nori domains: `*.tilework.tech` (Watchtower) and `*.noriskillsets.dev` (public/org registries, where `"public"` maps to the apex domain). `isPublicRegistryUrl({ url })` wraps `extractOrgId(url) === "public"` and is the canonical "is this the public registrar?" check (used by the profiles migration in `@/src/cli/profilesMigration.ts` to classify a profile's `.nori-version`).
 
-It also provides the reference-resolution helpers. `parseNamespacedPackage()` parses a `"org/package@version"` spec into `{ orgId, packageName, version }` and is the single chokepoint for org resolution: a bare, non-namespaced spec resolves to its optional `defaultOrg` argument when one is supplied, otherwise to `"public"` (the fallback is `orgId ?? defaultOrg ?? "public"`); an explicit `org/` or `public/` prefix is never overridden by `defaultOrg`. The companion `formatDefaultOrgNotice()` returns a one-line `Resolving "<name>" to "<org>/<name>"` info string when a bare name was routed to a non-public org (and `null` when an explicit namespace was given or the name resolved to public), which the registry-boundary commands log so the redirect is visible. Two distinct name builders then map the parsed reference to a name, and the split matters:
+It also provides the reference-resolution helpers. `parseNamespacedPackage()` parses a `"org/package@version"` spec into `{ orgId, packageName, version }` and is the single chokepoint for org resolution: a bare, non-namespaced spec resolves to its optional `defaultOrg` argument when one is supplied, otherwise to `"public"` (the fallback is `orgId ?? defaultOrg ?? "public"`); an explicit `org/` or `public/` prefix is never overridden by `defaultOrg`. The companion `formatDefaultOrgNotice()` returns a one-line `Resolving "<name>" to "<org>/<name>"` info string when a bare name was routed to a non-public org (and `null` when an explicit namespace was given or the name resolved to public), which the registry-boundary commands log so the redirect is visible.
 
-| Function | Returns | Meaning |
-|----------|---------|---------|
-| `namespacedName({ orgId, packageName })` | `packageName` for public, `<orgId>/<packageName>` for an org | **Reference name** handed to install/switch (bare for public). The persisted `activeSkillset` and `sks list`/`current` output instead use the fuller **namespaced identity** — see below |
-| `namespacedOnDiskName({ orgId, packageName })` | always `<orgId>/<packageName>` (so `public/<name>` for the public registrar) | **On-disk target** for downloads/installs — public packages land in the `public/` storage bucket |
+`namespacedName({ orgId, packageName })` then builds the **single fully-qualified `<orgId>/<packageName>` name** used everywhere below the CLI edge — user-facing display, the on-disk directory (relative to `profiles/`), and the stored identity. The public registrar is a normal namespace here, so a public package is `public/<name>`, never a bare `<name>`. This retires the former "public = bare" convention: the earlier bucket migration already made the local side use `public/foo` for list/current/switch/config/identity, so collapsing the old `namespacedOnDiskName` into `namespacedName` (they had become identical, and the separate `localSkillsetName` collapser was dead) makes the registry side match — one qualified name for reference, disk, and display.
 
-The name `namespacedName` produces is what `@/src/cli/commands/registry-install/registryInstall.ts` hands to install/switch, but `updateConfig` re-canonicalizes it to the namespaced identity on persist (`canonicalSkillsetName`, see `@/src/cli/config.ts`) and `sks list`/`current` render the identity via `skillsetIdentity` — so a public `foo` is handed in bare yet stored and shown as `public/foo`. The write/download path (`@/src/cli/commands/registry-download/registryDownload.ts`) uses `namespacedOnDiskName` to choose the bucketed directory. Reading an existing local skillset by bare name is handled by `resolveSkillsetDir` in `@/src/norijson/skillset.ts`, which re-resolves across the buckets. The `localSkillsetName` helper collapses a redundant `public/` prefix to a flat name.
+`namespacedName` is what `@/src/cli/commands/registry-install/registryInstall.ts` hands to install/switch and what `updateConfig` persists as `activeSkillset` (re-canonicalized via `canonicalSkillsetName`, see `@/src/cli/skillsetResolution.ts`); `sks list`/`current` render the on-disk identity via `skillsetIdentity` — so a public `foo` flows through and is stored and shown as `public/foo`. The write/download path (`@/src/cli/commands/registry-download/registryDownload.ts`) splits the same `namespacedName` on `/` to choose the bucketed directory (`profiles/public/<name>`). Reading an existing local skillset by bare name is still handled by `resolveSkillsetDir` in `@/src/norijson/skillset.ts`, which re-resolves across the buckets.
 
 **`fetch.ts`** initializes HTTP proxy support via undici's `EnvHttpProxyAgent` (call `initializeProxySupport()` once at startup), and defines four custom error classes: `NetworkError` (connectivity issues with proxy-aware messages), `ApiError` (HTTP status errors), `SkillCollisionError` (409 conflicts during skillset upload with per-skill conflict details), and `SubagentCollisionError` (409 conflicts during skillset upload with per-subagent conflict details, structurally mirroring `SkillCollisionError`). The `formatNetworkError()` function produces user-friendly messages that vary based on error code and proxy configuration.
